# 02-deep-learning-visual-lab-platform/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deep Learning Visual Lab API",
    "description": "Typed read and forward-pass API for the educational Deep Learning Visual Lab. Training is intentionally offline.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/cnn/predict": {
      "post": {
        "tags": [
          "CNN Feature Map Viewer"
        ],
        "summary": "Predict",
        "operationId": "predict_api_v1_cnn_predict_post",
        "parameters": [
          {
            "name": "sample_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CnnPredictionResource"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/api/v1/cnn/feature-maps": {
      "post": {
        "tags": [
          "CNN Feature Map Viewer"
        ],
        "summary": "Get Feature Maps",
        "operationId": "get_feature_maps_api_v1_cnn_feature_maps_post",
        "parameters": [
          {
            "name": "layer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Layer"
            }
          },
          {
            "name": "channels",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "maxItems": 12,
              "title": "Channels"
            }
          },
          {
            "name": "sample_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CnnFeatureMapsResource"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CnnFeatureMapResource": {
        "properties": {
          "layer": {
            "type": "string",
            "title": "Layer"
          },
          "layer_label": {
            "type": "string",
            "title": "Layer Label"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "channel": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Channel"
          },
          "tensor_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Tensor Shape"
          },
          "map_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Map Shape"
          },
          "raw_min": {
            "type": "number",
            "title": "Raw Min"
          },
          "raw_max": {
            "type": "number",
            "title": "Raw Max"
          },
          "raw_mean": {
            "type": "number",
            "title": "Raw Mean"
          },
          "raw_std": {
            "type": "number",
            "minimum": 0.0,
            "title": "Raw Std"
          },
          "normalization": {
            "type": "string",
            "title": "Normalization"
          },
          "display_scale": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Display Scale"
          },
          "values": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "minItems": 1,
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "layer",
          "layer_label",
          "operation",
          "channel",
          "tensor_shape",
          "map_shape",
          "raw_min",
          "raw_max",
          "raw_mean",
          "raw_std",
          "normalization",
          "display_scale",
          "values"
        ],
        "title": "CnnFeatureMapResource"
      },
      "CnnFeatureMapsResource": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "input": {
            "$ref": "#/components/schemas/CnnInputResource"
          },
          "prediction": {
            "$ref": "#/components/schemas/CnnPredictionSummaryResource"
          },
          "representation": {
            "$ref": "#/components/schemas/CnnRepresentationResource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "module",
          "model_version",
          "input",
          "prediction",
          "representation"
        ],
        "title": "CnnFeatureMapsResource"
      },
      "CnnInputResource": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Id"
          },
          "source_index": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Index"
          },
          "registered_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registered Label"
          },
          "original_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Original Shape"
          },
          "tensor_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Tensor Shape"
          },
          "preprocessing": {
            "$ref": "#/components/schemas/CnnPreprocessingResource"
          },
          "image_data_uri": {
            "type": "string",
            "pattern": "^data:image/png;base64,",
            "title": "Image Data Uri"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "original_shape",
          "tensor_shape",
          "preprocessing",
          "image_data_uri"
        ],
        "title": "CnnInputResource"
      },
      "CnnLayerResource": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "channels": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Channels"
          },
          "tensor_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Tensor Shape"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "operation",
          "channels",
          "tensor_shape"
        ],
        "title": "CnnLayerResource"
      },
...
      "CnnPredictionResource": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "input": {
            "$ref": "#/components/schemas/CnnInputResource"
          },
          "prediction": {
            "$ref": "#/components/schemas/CnnPredictionSummaryResource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "module",
          "model_version",
          "input",
          "prediction"
        ],
        "title": "CnnPredictionResource"
      },
      "CnnPredictionSummaryResource": {
        "properties": {
          "predicted_index": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 0.0,
            "title": "Predicted Index"
          },
          "predicted_class": {
            "type": "string",
            "title": "Predicted Class"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "probabilities": {
            "items": {
              "$ref": "#/components/schemas/CnnProbabilityResource"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 10,
            "title": "Probabilities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "predicted_index",
          "predicted_class",
          "confidence",
          "probabilities"
        ],
        "title": "CnnPredictionSummaryResource"
      },
      "CnnPreprocessingResource": {
        "properties": {
          "color_mode": {
            "type": "string",
            "title": "Color Mode"
          },
          "image_size": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "prefixItems": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              }
            ],
            "title": "Image Size"
          },
          "value_range": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              }
            ],
            "title": "Value Range"
          },
          "resize_policy": {
            "type": "string",
            "title": "Resize Policy"
          },
          "normalization": {
            "type": "string",
            "title": "Normalization"
          },
          "mean": {
            "type": "number",
            "title": "Mean"
          },
          "std": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Std"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "color_mode",
          "image_size",
          "value_range",
          "resize_policy",
          "normalization",
          "mean",
          "std"
        ],
        "title": "CnnPreprocessingResource"
      },
      "CnnProbabilityResource": {
        "properties": {
          "class_index": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 0.0,
            "title": "Class Index"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probability"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "class_index",
          "class_name",
          "probability"
        ],
        "title": "CnnProbabilityResource"
      },
      "CnnRepresentationResource": {
        "properties": {
          "layer": {
            "$ref": "#/components/schemas/CnnLayerResource"
          },
          "activation_tensor_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Activation Tensor Shape"
          },
          "maps": {
            "items": {
              "$ref": "#/components/schemas/CnnFeatureMapResource"
            },
            "type": "array",
            "minItems": 1,
            "title": "Maps"
          },
          "comparison_rule": {
            "type": "string",
            "title": "Comparison Rule"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "layer",
          "activation_tensor_shape",
          "maps",
          "comparison_rule"
        ],
        "title": "CnnRepresentationResource"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorResource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "ErrorItemResource": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorItemResource"
      },
      "ErrorResource": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorItemResource"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResource"
      },
...
    }
  }
}